# apps/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Feedback Flow API",
    "version": "1.7.0"
  },
  "paths": {
...
    "/api/refund": {
      "post": {
        "summary": "Record refund",
        "description": "Records a refund for a purchase. Requires write permission.",
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearerAuth": [
              "write:api"
            ]
          },
          {
            "oauth2": [
              "write:api"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Refund successfully recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or missing required fields"
          }
        }
      }
    },
    "/api/refund/{id}": {
      "get": {
        "summary": "Get refund info",
        "description": "Returns information about a specific refund. Requires read permission.",
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearerAuth": [
              "read:api"
            ]
          },
          {
            "oauth2": [
              "read:api"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Purchase ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved refund info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Refund"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Refund not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RefundCreateRequest": {
        "type": "object",
        "required": [
          "date",
          "purchase",
          "refundDate",
          "amount",
          "transactionId"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "purchase": {
            "type": "string"
          },
          "refundDate": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "Refund": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "purchase": {
            "type": "string"
          },
          "refundDate": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "type": "number"
          },
          "transactionId": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "description": "Auth0 OAuth2 Authorization Code Flow",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sctg.eu.auth0.com/authorize",
            "tokenUrl": "https://sctg.eu.auth0.com/oauth/token",
            "scopes": {
              "read:api": "Read Api",
              "write:api": "Write Api",
              "admin:api": "Admin Api",
              "backup:api": "Backup Api",
              "search:api": "Search Api",
              "auth0:admin:api": "Auth0 Admin Api",
              "openid": "Openid",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "oauth2": []
    }
  ]
}